www/hs-feed: import hs-feed-1.3.2.1
Interfacing with RSS (v 0.9x, 2.x, 1.0) + Atom feeds.

* Parsers
* Constructors
* Rendering
* Querying

To help working with the multiple feed formats we've ended up with this set
of modules providing parsers, printers and some utility code for querying
and just generally working with a concrete representation of feeds in
Haskell.

For basic reading and editing of feeds, consult the documentation of the
Text.Feed.* hierarchy.