py-joblib: updated to 1.3.2
Release 1.3.2 -- 2023/08/08

Fix a regression in joblib.Parallel introduced in 1.3.0 where explicitly setting n_jobs=None was not interpreted as "unset".
Fix a regression in joblib.Parallel introduced in 1.3.0 where joblib.Parallel logging methods exposed from inheritance to joblib.Logger didn't work because of missing logger initialization.
Various maintenance updates to the doc, the ci and the test.

Release 1.3.1 -- 2023/06/29

Fix compatibility with python 3.7 by vendor loky 3.4.1 which is compatible with this version.

Release 1.3.0 -- 2023/06/28

Ensure native byte order for memmap arrays in joblib.load.
Add ability to change default Parallel backend in tests by setting the JOBLIB_TESTS_DEFAULT_PARALLEL_BACKEND environment variable.
Fix temporary folder creation in joblib.Parallel on Linux subsystems on Windows which do have /dev/shm but don't have the os.statvfs function
Drop runtime dependency on distutils. distutils is going away in Python 3.12 and is deprecated from Python 3.10 onwards. This import was kept around to avoid breaking scikit-learn, however it's now been long enough since scikit-learn deployed a fixed (verion 1.1 was released in May 2022) that it should be safe to remove this.
A warning is raised when a pickling error occurs during caching operations. In version 1.5, this warning will be turned into an error. For all other errors, a new warning has been introduced: joblib.memory.CacheWarning.
Avoid (module, name) collisions when caching nested functions. This fix changes the module name of nested functions, invalidating caches from previous versions of Joblib.
Add cache_validation_callback in :meth:`joblib.Memory.cache`, to allow custom cache invalidation based on the metadata of the function call.
Add a return_as parameter for Parallel, that enables consuming results asynchronously.
Improve the behavior of joblib for n_jobs=1, with simplified tracebacks and more efficient running time.
Add the parallel_config context manager to allow for more fine-grained control over the backend configuration. It should be used in place of the parallel_backend context manager. In particular, it has the advantage of not requiring to set a specific backend in the context manager.
Add items_limit and age_limit in :meth:`joblib.Memory.reduce_size` to make it easy to limit the number of items and remove items that have not been accessed for a long time in the cache.
Deprecate bytes_limit in Memory as this is not automatically enforced, the limit can be directly passed to :meth:`joblib.Memory.reduce_size` which needs to be called to actually enforce the limit.
Vendor loky 3.4.0 which includes various fixes.
Various updates to the documentation and to benchmarking tools.
Move project metadata to pyproject.toml.
Add more tests to improve python nogil support.