py-idna: updated to 3.6
3.6 (2023-11-25)

Fix regression to include tests in source distribution.

3.5 (2023-11-24)

Update to Unicode 15.1.0
String codec name is now "idna2008" as overriding the system codec "idna" was not working.
Fix typing error for codec encoding
"setup.cfg" has been added for this release due to some downstream lack of adherence to PEP 517. Should be removed in a future release so please prepare accordingly.
Removed reliance on a symlink for the "idna-data" tool to comport with PEP 517 and the Python Packaging User Guide for sdist archives.
Added security reporting protocol for project